requirements.txt
# Core ML/Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Deep Learning (for LSTM variants)
tensorflow>=2.15.0

# Model Explainability
shap>=0.44.0

# MLOps & Experiment Tracking
mlflow>=2.9.0

# LangGraph for agent pipelines
langgraph>=0.0.40
langchain>=0.1.0

# Data Validation
great-expectations>=0.18.0

# API & Web
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.66.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
